fix sign_with_secret_key
